pyproject.toml
[tool.poetry]
name = "Similarity_Graph"
version = "0.1.0"
description = ""
authors = ["Vahid Shahrivari"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
pymilvus = "^2.1.3"
jupyter = "^1.0.0"
matplotlib = "^3.6.1"
sklearn = "^0.0"
networkx = "^2.8.7"
imbalanced-learn = "^0.9.1"
dataframe-image = "^0.1.3"
seaborn = "^0.12.1"
hydra-core = "^1.2.0"
tqdm = "^4.64.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
